refactor: domain 패키지 내 클래스들을 계층화
Lotto 관련 클래스들은 LottoGame의 하위 클래스로 정의하여, sub 패키지 밑으로 이동
