+++ src/main/java/edu/nextstep/lotto/domain/sub/LottoNumber.java
@@ -1,4 +1,4 @@
-package edu.nextstep.lotto.domain;
+package edu.nextstep.lotto.domain.sub;
 
 import java.util.Objects;
