+++ src/test/java/edu/nextstep/lotto/domain/LottoNumberTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import edu.nextstep.lotto.domain.sub.LottoNumber;
+
 @DisplayName("LottoNumber: 로또 번호 각각의 원시값을 포장한 클래스")
 class LottoNumberTest {
 	@DisplayName("생성자: LottoNumber 객체를 정상적으로 생성하여 리턴함")
